git-svn-id: https://forge.ill.fr/svn/crysfml@103 06508225-4b59-0410-a2a9-fdab55894d57

+++ Src/CFML/CFML_Atom_Mod.f90
@@ -117,7 +117,7 @@
     !!----    real(kind=sp),dimension(6)              :: U_std         ! Standar_Deviations of U"s
     !!----    real(kind=sp)                           :: Ueq           ! Uequiv
     !!----    real(kind=sp),dimension(6)              :: mU            !
-    !!----    real(kind=sp),dimension(6)              :: lU            !
+    !!----    integer,dimension(6)                    :: lU            !
     !!----    real(kind=sp)                           :: Charge        ! Charge
     !!----    real(kind=sp)                           :: Moment        ! Moment
     !!----    integer, dimension(5)                   :: Ind           ! Index for different purposes
